+++ CHANGELOG.md
@@ -1,5 +1,16 @@
 # Changelog
 
+## [1.2.0] - Unreleased
+
+### Fixed
+
+- `resolved.CivisFuture` now gets the status of the platform job directly and updates the state of the future.
+
+### Added
+
+- `value.CivisFuture` now prints error logs from failed jobs, throws a catcheable
+error, and fetches the job logs automatically.
+
 ## [1.1.1] - 2017-11-20
 
 ### Fixed